refactor: split content js into content/*.js
