+++ src/js/content/constants.js
@@ -1,0 +1,1 @@
+export const THREE_SECONDS = 3000
