+++ src/js/content/utils.js
@@ -1,0 +1,1 @@
+const hideElem = htmlElem => (htmlElem.style.display = "none")
